+++ src/components/uavs/UAVOperationsButtonGroup.jsx
@@ -1,8 +1,9 @@
 import isEmpty from 'lodash-es/isEmpty';
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, { useCallback, useState } from 'react';
 import { withTranslation } from 'react-i18next';
 import { connect } from 'react-redux';
+import { useInterval } from 'react-use';
 import { bindActionCreators } from '@reduxjs/toolkit';
 
 import Button from '@material-ui/core/Button';
@@ -27,6 +28,7 @@ import Colors from '~/components/colors';
 import ToolbarDivider from '~/components/ToolbarDivider';
 import Bolt from '~/icons/Bolt';
 
+import { UAV_SIGNAL_DURATION } from '~/features/settings/constants';
 import {
   requestRemovalOfUAVsByIds,
   requestRemovalOfUAVsMarkedAsGone,
@@ -89,16 +91,33 @@ const UAVOperationsButtonGroup = ({
     dispatch
   );
 
+  const [keepFlashing, setKeepFlashing] = useState(false);
+  const flashLightsButtonOnClick = useCallback(
+    (event) => {
+      if (keepFlashing) {
+        setKeepFlashing(false);
+      } else if (event.shiftKey) {
+        setKeepFlashing(true);
+        flashLight();
+      } else {
+        flashLight();
+      }
+    },
+    [flashLight, keepFlashing, setKeepFlashing]
+  );
+
+  useInterval(flashLight, keepFlashing ? UAV_SIGNAL_DURATION * 1000 : null);
+
   const fontSize = size === 'small' ? 'small' : 'medium';
   const iconSize = size;
 
   const flashLightsButton =
     size === 'small' ? (
       <Button
-        startIcon={<WbSunny />}
+        startIcon={<WbSunny color={keepFlashing ? 'primary' : undefined} />}
         disabled={isSelectionEmpty}
         size={iconSize}
-        onClick={flashLight}
+        onClick={flashLightsButtonOnClick}
       >
         {t('UAVOpButtonGrp.flashLights')}
       </Button>
@@ -107,9 +126,12 @@ const UAVOperationsButtonGroup = ({
         <IconButton
           disabled={isSelectionEmpty}
           size={iconSize}
-          onClick={flashLight}
+          onClick={flashLightsButtonOnClick}
         >
-          <WbSunny fontSize={fontSize} />
+          <WbSunny
+            fontSize={fontSize}
+            color={keepFlashing ? 'primary' : undefined}
+          />
         </IconButton>
       </Tooltip>
     );